Optimized + Updated class name
Updated the selector for the header element (since Spotify changed a bunch of class names AGAIN).
Made the css var local to the header so the entire document isn't being modified every second (which also fixes the issues which arise when trying to use the devtools inspector while this extension is enabled).