Optional chaining added
Changed `style` to `qt_style` in case any other scripts initialize a var with the same name.
Added optional chaining when grabbing `Spicetify.Queue.nextTracks` and when attempting to update the css var so errors aren't spit out when not on the queue page (the errors weren't causing problems, but cleaner to just prevent them anyway).